requirements.txt
streamlit
ipykernel
openai
python-dotenv
heyoo
pygsheets
pandas

